% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes.R
\name{list_solar_prettysignal_examples}
\alias{list_solar_prettysignal_examples}
\title{List available example scripts}
\usage{
list_solar_prettysignal_examples()
}
\description{
This function lists the example scripts available in the package.
}
