% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes.R
\name{run_solar_prettysignal_example}
\alias{run_solar_prettysignal_example}
\title{Run an example script}
\usage{
run_solar_prettysignal_example(example_name)
}
\arguments{
\item{example_name}{The name of the example script to run (without extension).}
}
\description{
This function runs a specified example script.
}
