% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes.R
\name{solar_prettysignal}
\alias{solar_prettysignal}
\title{Solar Pretty Signal}
\usage{
solar_prettysignal(dados_to_be_treated, midnight_is_first_hour)
}
\arguments{
\item{dados_to_be_treated}{DataFrame with the values to be processed. This DataFrame must have the columns \emph{Data}, \emph{Year}, \emph{Month}, \emph{Day}, and \emph{Hour}. The \emph{Data} column should contain the values to be processed, which can be measured data of photovoltaic generation or solar irradiation. 24 values per day must be provided.}

\item{midnight_is_first_hour}{Boolean indicating whether \emph{midnight} is the first hour of the day.}
}
\value{
Data frame with the \emph{original data} and the \emph{filtered data}.
}
\description{
This function receives measured data of photovoltaic generation or solar irradiation, applies filtering and processing and returns the filtered data.
}
